+++ ChangeLog.md
@@ -8,6 +8,7 @@
 - ADDED: #4087 The installer packages are back (powered by VMware InstallBuilder), with proper "Start Menu" entries and SQLite file associations - across all 3 platforms!
 - ADDED: #3291 Filtering in View window (thanks to @electrickite).
 - ADDED: #4103 Windows 32-bit binaries, alongside with 64-bit. With version 3.3.0 Windows binaries were upgraded to 64 bit, but it turns out many people still need 32 bit binaries.
+- ADDED: #3715 Dedicated hotkeys to execute either single query under cursor or all queries in editor, regardless of default execution mode configured. Default hotkeys: Ctrl+F9 (or Cmd+F9) for single query, Shift+F9 for all queries.
 - CHANGE: #3272 Named function parameters of Custom SQL functions are now passed to script code as named variables.
 - CHANGE: #3337 QtScript (deprecated module) usage migrated to QML module, using QJSEngine, the EcmaScript compliant implementation. Also changed plugin language name from QtScript to JavaScript and icon from Qt icon to JS icon.
 - CHANGE: #2963 Application state (session) is saved (apart from normal application exit) whenever the state changes and also during critical application crash.